API: Move ipm out of the epics folder
